+++ CHANGES.md
@@ -1,6 +1,7 @@
 # 1.1 (TBD)
 * Pressing C to adjust pivot distance now ignores backfaces when backface culling is enabled in viewport settings.
 * Added camera roll mouse action. Can press R to reset roll.
+* Mouse strafing now remains active while at least one mouse button is still pressed, regardless of whether strafing key was released.
 
 # 1.0 (2019-10-01)
 * Initial release