+++ CHANGES.md
@@ -1,9 +1,10 @@
-# 1.1 (TBD)
+# 2.0 (2020-12-20)
 * Pressing C to adjust pivot distance now ignores backfaces when backface culling is enabled in viewport settings.
 * Added camera roll mouse action. Can press R to reset roll.
 * Mouse strafing now remains active while at least one mouse button is still pressed, regardless of whether strafing key was released.
-* Addon now cooperates with "Lock camera to view" setting.
+* Addon now cooperates with "Lock camera to view" and "Local camera" settings.
 * Improved ergonomics of mouse acceleration.
+* `package.bat` now creates Linux-compatible archives.
 
 # 1.0 (2019-10-01)
 * Initial release